Extend parse for parsing newick/nexus input, and add format keyword to write.
