+++ test/test_newick.jl
@@ -68,17 +68,17 @@ using Test
             tree2 = open(f -> parsenewick(f, TreeType), "t1.newick")
             @test nleaves(tree) == nleaves(tree2) == 507
             @test ntrees(tree) == ntrees(tree2) == 1
-            @test_nowarn Phylo.write("t1.trees", tree, Nexus())
+            @test_nowarn Phylo.write("t1.trees", tree, format = Nexus())
             tree2x = open(f -> parsenexus(f, TreeType), "t1.trees")
             @test nleaves(tree) == nleaves(tree2x) == 507
             @test ntrees(tree) == ntrees(tree2x) == 1
             @test Set(getnodenames(tree)) == Set(getnodenames(tree2))
             @test Set(getleafnames(tree)) == Set(getleafnames(tree2)) == Set(getleafnames(tree2x))
             names = getnodenames(tree2)
-            Phylo.write("t2.newick", tree2, Newick(Dict(names .=> 1:nnodes(tree2))))
+            Phylo.write("t2.newick", tree2, format = Newick(Dict(names .=> 1:nnodes(tree2))))
             tree3 = open(f -> parsenewick(f, TreeType), "t2.newick")
             @test Set(getnodenames(tree3)) == Set(string.(1:nnodes(tree3)))
-            Phylo.write("t3.newick", tree3, Newick(Dict(string.(1:nnodes(tree3)) .=> names)))
+            Phylo.write("t3.newick", tree3, format = Newick(Dict(string.(1:nnodes(tree3)) .=> names)))
             tree4 = open(f -> parsenewick(f, TreeType), "t3.newick")
             @test Set(getnodenames(tree)) == Set(getnodenames(tree4))
             @test Set(getleafnames(tree)) == Set(getleafnames(tree4))
